Move slurmrestd inference to worker utils and fix task manager
