+++ beeflow/common/worker/utils.py
@@ -1,6 +1,8 @@
 """Worker utility functions."""
 
 import subprocess
+from packaging.version import Version
+
 from beeflow.common.worker.worker import WorkerError
 from beeflow.common import log as bee_logging
 
@@ -32,3 +34,20 @@ def parse_key_val(pair):
     """Parse the key-value pair separated by '='."""
     i = pair.find('=')
     return (pair[:i], pair[i + 1:])
+
+def get_slurmrestd_version():
+    """Get the newest slurmrestd version."""
+    resp = subprocess.run(["slurmrestd", "-s", "list"], check=True, stderr=subprocess.PIPE,
+                          text=True).stderr
+    resp = resp.split("\n")
+    # Confirm slurmrestd format is the same
+    # If the slurmrestd list outputs has changed potentially something else has broken
+    if "Possible OpenAPI plugins" not in resp[0]:
+        print("Slurmrestd OpenAPI format has changed and things may break")
+    api_versions = [line.split('/')[1] for line in resp[1:] if re.search(r"openapi/v\d+\.\d+\.\d+",
+                                                                         line)]
+    # Sort the versions and grab the newest one
+    newest_api = sorted(api_versions, key=Version, reverse=True)[0]
+    print(f"Inferred slurmrestd version: {newest_api}")
+    return newest_api
+
